finished refactoring playlist cqrs and controller :D
