--- RussianSpotify.API/RussianSpotify.API.Contracts/Requests/Music/GetPlaylistsByFilter/GetPlaylistsByFilterResponse.cs
@@ -1,24 +1,0 @@
-﻿namespace RussianSpotify.Contracts.Requests.Music.GetPlaylistsByFilter;
-
-public class GetPlaylistsByFilterResponse
-{
-    /// <summary>
-    /// Название альбома
-    /// </summary>
-    public string PlaylistName { get; set; } = default!;
-
-    /// <summary>
-    /// ИД картинки
-    /// </summary>
-    public Guid? ImageId { get; set; }
-
-    /// <summary>
-    /// Автор
-    /// </summary>
-    public string? AuthorName { get; set; }
-
-    /// <summary>
-    /// Дата релиза
-    /// </summary>
-    public DateTime ReleaseDate { get; set; }
-}
