+++ RussianSpotify.API/RussianSpotify.API.Contracts/Requests/Playlist/GetPlaylistsByFilter/GetPlaylistsByFilterResponse.cs
@@ -1,0 +1,24 @@
+﻿namespace RussianSpotify.Contracts.Requests.Playlist.GetPlaylistsByFilter;
+
+public class GetPlaylistsByFilterResponse
+{
+    /// <summary>
+    /// Название альбома
+    /// </summary>
+    public string PlaylistName { get; set; } = default!;
+
+    /// <summary>
+    /// ИД картинки
+    /// </summary>
+    public Guid? ImageId { get; set; }
+
+    /// <summary>
+    /// Автор
+    /// </summary>
+    public string? AuthorName { get; set; }
+
+    /// <summary>
+    /// Дата релиза
+    /// </summary>
+    public DateTime ReleaseDate { get; set; }
+}
