+++ RussianSpotify.API/RussianSpotify.API.Contracts/RussianSpotify.API.Contracts.csproj
@@ -18,6 +18,10 @@
 
 
 
+    <ItemGroup>
+      <Folder Include="Requests\Playlist\" />
+    </ItemGroup>
+
     </ItemGroup>
