+++ RussianSpotify.API/RussianSpotify.API.Core/Requests/Playlist/GetPlaylistsByFilter/GetPlaylistsByFilterQuery.cs
@@ -1,0 +1,16 @@
+﻿using MediatR;
+using RussianSpotify.Contracts.Requests.Playlist.GetPlaylistsByFilter;
+
+namespace RussianSpotify.API.Core.Requests.Playlist.GetPlaylistsByFilter;
+
+/// <summary>
+/// Запрос на получение плейлистов по фильтру
+/// </summary>
+public class GetPlaylistsByFilterQuery
+    : GetPlaylistsByFilterRequest, IRequest<List<GetPlaylistsByFilterResponse>>
+{
+    public GetPlaylistsByFilterQuery(GetPlaylistsByFilterRequest request)
+        : base(request)
+    {
+    }
+}
