+++ RussianSpotify.API/RussianSpotify.API.Core/RussianSpotify.API.Core.csproj
@@ -53,4 +53,8 @@
 
 
 
+    <ItemGroup>
+      <Folder Include="Requests\Playlist\" />
+    </ItemGroup>
+
 </Project>
