+++ RussianSpotify.API/RussianSpotify.API.WEB/Controllers/PlaylistController.cs
@@ -1,14 +1,16 @@
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
-using RussianSpotify.API.Core.Requests.Music.GetAllFavouriteAlbumAndPlaylist;
-using RussianSpotify.API.Core.Requests.Music.GetFavouritePlaylistById;
-using RussianSpotify.API.Core.Requests.Music.PostAddPlaylistToFavourite;
-using RussianSpotify.API.Core.Requests.Music.PostCreatePlaylist;
-using RussianSpotify.API.Core.Requests.Music.PutPlaylist;
-using RussianSpotify.Contracts.Requests.Music.GetAllFavouriteAlbumAndPlaylist;
-using RussianSpotify.Contracts.Requests.Music.GetFavouritePlaylistById;
-using RussianSpotify.Contracts.Requests.Music.PostCreatePlaylist;
-using RussianSpotify.Contracts.Requests.Music.PutPlaylist;
+using RussianSpotify.API.Core.Requests.Playlist.GetAllFavouriteAlbumAndPlaylist;
+using RussianSpotify.API.Core.Requests.Playlist.GetFavouritePlaylistById;
+using RussianSpotify.API.Core.Requests.Playlist.GetPlaylistsByFilter;
+using RussianSpotify.API.Core.Requests.Playlist.PostAddPlaylistToFavourite;
+using RussianSpotify.API.Core.Requests.Playlist.PostCreatePlaylist;
+using RussianSpotify.API.Core.Requests.Playlist.PutPlaylist;
+using RussianSpotify.Contracts.Requests.Playlist.GetAllFavouriteAlbumAndPlaylist;
+using RussianSpotify.Contracts.Requests.Playlist.GetFavouritePlaylistById;
+using RussianSpotify.Contracts.Requests.Playlist.GetPlaylistsByFilter;
+using RussianSpotify.Contracts.Requests.Playlist.PostCreatePlaylist;
+using RussianSpotify.Contracts.Requests.Playlist.PutPlaylist;
 
 namespace RussianSpotify.API.WEB.Controllers;
 
@@ -26,6 +28,25 @@ public class PlaylistController : ControllerBase
     public PlaylistController(IMediator mediator)
     {
         _mediator = mediator;
+    }
+    
+    /// <summary>
+    /// Получить альбомы по фильтру(Доступные фильтры: AuthorPlaylists)
+    /// </summary>
+    /// <param name="request">GetPlaylistsByFilterRequest(Название фильтра,
+    /// значение фильтра, страница, кол-во альбомов на странице)</param>
+    /// <param name="cancellationToken">Токен отмены</param>
+    /// <returns>Список GetPlaylistsByFilterResponse альбомы по фильтру</returns>
+    [HttpGet("GetPlaylistsByFilter")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    public async Task<List<GetPlaylistsByFilterResponse>> GetPlaylistsByFilter(
+        [FromQuery] GetPlaylistsByFilterRequest request,
+        CancellationToken cancellationToken)
+    {
+        var query = new GetPlaylistsByFilterQuery(request);
+        return await _mediator.Send(query, cancellationToken);
     }
 
     /// <summary>
